Separate profile section into another file from settings and transferred into home page
