+++ lib/pages/home_page/home_layout.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flilipino_food_app/pages/favorite/favorite_page.dart';
 import 'package:flilipino_food_app/pages/home_page/home_page.dart';
-import 'package:flilipino_food_app/pages/home_page/home_widgets/profile_section.dart';
+import 'package:flilipino_food_app/pages/settings/settings_page.dart';
 import 'package:flilipino_food_app/pages/home_page/home_widgets/search_recipe.dart';
 import 'package:flutter/material.dart';
 
@@ -42,13 +42,13 @@ class _HomeLayoutState extends State<HomeLayout> {
         destinations: const [
           NavigationDestination(icon: Icon(Icons.home), label: "Home"),
           NavigationDestination(icon: Icon(Icons.bookmarks), label: "Saved"),
-          NavigationDestination(icon: Icon(Icons.person), label: "Profile")
+          NavigationDestination(icon: Icon(Icons.settings), label: "Settings")
         ],
       ),
       body: [
         const HomePage(),
         const FavoritePage(),
-        const ProfileSection(),
+        const SettingsPage(),
       ][_currentPageIndex],
     );
   }
